+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - linter and formater
 - example view with Amazon IVS player
 - quick & nimble test framework
+- Amazon IVS adapter & listeners
 
 ### Changed
 - updated example app